+++ src/features/mm_facing_feature.h
@@ -11,7 +11,7 @@ using namespace godot;
 class MMFacingFeature : public MMFeature {
     GDCLASS(MMFacingFeature, MMFeature)
 
-    virtual size_t get_dimension_count() const override;
+    virtual uint32_t get_dimension_count() const override;
 
     virtual void setup_skeleton(const MMAnimationPlayer* p_player, const Skeleton3D* p_skeleton) override;
